Bug 138277 EFS Zip File System Example "Collapse Zip File" Error

+++ platform-ui/plugins/org.eclipse.ui.examples.filesystem/src/org/eclipse/ui/examples/filesystem/CollapseZipAction.java
@@ -1,16 +1,21 @@
 package org.eclipse.ui.examples.filesystem;
 
 import java.net.URI;
-import java.net.URISyntaxException;
 import org.eclipse.core.filesystem.EFS;
-import org.eclipse.core.resources.*;
+import org.eclipse.core.filesystem.IFileStore;
+import org.eclipse.core.filesystem.URIUtil;
+import org.eclipse.core.resources.IFile;
+import org.eclipse.core.resources.IFolder;
+import org.eclipse.core.resources.IResource;
 import org.eclipse.core.runtime.Path;
 import org.eclipse.jface.action.IAction;
 import org.eclipse.jface.dialogs.MessageDialog;
 import org.eclipse.jface.viewers.ISelection;
 import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.swt.widgets.Shell;
-import org.eclipse.ui.*;
+import org.eclipse.ui.IActionDelegate;
+import org.eclipse.ui.IObjectActionDelegate;
+import org.eclipse.ui.IWorkbenchPart;
 
 public class CollapseZipAction implements IObjectActionDelegate {
 
@@ -26,14 +31,26 @@ public class CollapseZipAction implements IObjectActionDelegate {
 
 	private void collapseZip(IFolder folder) {
 		try {
-			folder.delete(IResource.NONE, null);
-			folder.getParent().refreshLocal(IResource.DEPTH_INFINITE, null);
+			URI zipURI = new URI(folder.getLocationURI().getQuery());
+			//check if the zip file is physically stored below the folder in the workspace
+			IFileStore parentStore = EFS.getStore(folder.getParent().getLocationURI());
+			URI childURI = parentStore.getChild(folder.getName()).toURI();
+			if (URIUtil.equals(zipURI, childURI)) {
+				//the zip file is in the workspace so just delete the link 
+				// and refresh the parent to create the resource
+				folder.delete(IResource.NONE, null);
+				folder.getParent().refreshLocal(IResource.DEPTH_INFINITE, null);
+			} else {
+				//otherwise the zip file must be a linked resource
+				IFile file = folder.getParent().getFile(new Path(folder.getName()));
+				file.createLink(zipURI, IResource.REPLACE, null);
+			}
 		} catch (Exception e) {
 			MessageDialog.openError(getShell(), "Error", "Error opening zip file");
 			e.printStackTrace();
 		}
 	}
-	
+
 	private Shell getShell() {
 		return targetPart.getSite().getShell();
 	}
